Refactor rules to enable external configuration
This is work in progress but I'd like to get it into master so others can build off it rather than diverging.

Rules are now configured by ```rulesets/default.json``` which defines which rules to run for which language.

Rules can point at a module specific to that module (e.g. the license-detectable-by-licensee) or can "subclass" an existing rule (e.g. license-file-exists:file-existence).

This PR also brings in the use of the Standard linter as recommended by @bkeepers in #10.

The work I'll do in a subsequent PR is splitting up ```index.js``` into independent functions for ruleset loading and execution as well as enabling users to override the ruleset used and to support ruleset inheritence.