Add a new ruleset with Git-related checks
The git ruleset contains two new rules that search the commit messages
and the commits themselves for configurable blacklisted words. These
words can in fact be extended regular expressions. This is a new ruleset
instead of adding the rules to the default ruleset as these checks can be
a bit time consuming, depending on the size of the Git history.