PR#29: Check Code of Conduct for Email Address
This ensures there is an email address in the code of conduct file.

This also extends the file-contents functionality to have a
human-readable contents for those cases where the contents you are
searching for is not human-readable, like a regexp.

Updated README to reflect the changes.

Fixed an overly-strict git grep commits test.